fix: add .toByte() to ALL hex values in byteArrayOf

+++ app/src/main/java/com/example/bluetoothkeyboard/Constants.kt
@@ -11,46 +11,46 @@ object Constants {
     // 键盘报告描述符
     // 标准 USB HID 键盘描述符
     val SDP_RECORD = byteArrayOf(
-        0x05, 0x01.toByte(), // Usage Page (Generic Desktop)
-        0x09, 0x06,          // Usage (Keyboard)
-        0xA1.toByte(), 0x01, // Collection (Application)
-        0x05, 0x07,          //   Usage Page (Key Codes)
-        0x19, 0xE0.toByte(), //   Usage Minimum (224)
-        0x29, 0xE7.toByte(), //   Usage Maximum (231)
-        0x15, 0x00,          //   Logical Minimum (0)
-        0x25, 0x01,          //   Logical Maximum (1)
-        0x75, 0x01,          //   Report Size (1)
-        0x95, 0x08,          //   Report Count (8)
-        0x81.toByte(), 0x02, //   Input (Data, Variable, Absolute)
-        0x95, 0x01,          //   Report Count (1)
-        0x75, 0x08,          //   Report Size (8)
-        0x81.toByte(), 0x01, //   Input (Constant)
-        0x95, 0x05,          //   Report Count (5)
-        0x75, 0x01,          //   Report Size (1)
-        0x05, 0x08,          //   Usage Page (Page# for LEDs)
-        0x19, 0x01,          //   Usage Minimum (1)
-        0x29, 0x05,          //   Usage Maximum (5)
-        0x91.toByte(), 0x02, //   Output (Data, Variable, Absolute)
-        0x95, 0x01,          //   Report Count (1)
-        0x75, 0x03,          //   Report Size (3)
-        0x91.toByte(), 0x01, //   Output (Constant)
-        0x95, 0x06,          //   Report Count (6)
-        0x75, 0x08,          //   Report Size (8)
-        0x15, 0x00,          //   Logical Minimum (0)
-        0x25, 0x65,          //   Logical Maximum (101)
-        0x05, 0x07,          //   Usage Page (Key Codes)
-        0x19, 0x00,          //   Usage Minimum (0)
-        0x29, 0x65,          //   Usage Maximum (101)
-        0x81.toByte(), 0x00, //   Input (Data, Array)
-        0xC0.toByte()        // End Collection
+        0x05.toByte(), 0x01.toByte(),
+        0x09.toByte(), 0x06.toByte(),
+        0xA1.toByte(), 0x01.toByte(),
+        0x05.toByte(), 0x07.toByte(),
+        0x19.toByte(), 0xE0.toByte(),
+        0x29.toByte(), 0xE7.toByte(),
+        0x15.toByte(), 0x00.toByte(),
+        0x25.toByte(), 0x01.toByte(),
+        0x75.toByte(), 0x01.toByte(),
+        0x95.toByte(), 0x08.toByte(),
+        0x81.toByte(), 0x02.toByte(),
+        0x95.toByte(), 0x01.toByte(),
+        0x75.toByte(), 0x08.toByte(),
+        0x81.toByte(), 0x01.toByte(),
+        0x95.toByte(), 0x05.toByte(),
+        0x75.toByte(), 0x01.toByte(),
+        0x05.toByte(), 0x08.toByte(),
+        0x19.toByte(), 0x01.toByte(),
+        0x29.toByte(), 0x05.toByte(),
+        0x91.toByte(), 0x02.toByte(),
+        0x95.toByte(), 0x01.toByte(),
+        0x75.toByte(), 0x03.toByte(),
+        0x91.toByte(), 0x01.toByte(),
+        0x95.toByte(), 0x06.toByte(),
+        0x75.toByte(), 0x08.toByte(),
+        0x15.toByte(), 0x00.toByte(),
+        0x25.toByte(), 0x65.toByte(),
+        0x05.toByte(), 0x07.toByte(),
+        0x19.toByte(), 0x00.toByte(),
+        0x29.toByte(), 0x65.toByte(),
+        0x81.toByte(), 0x00.toByte(),
+        0xC0.toByte()
     )
 
     // QoS 参数
     val QOS_OUT = byteArrayOf(
-        0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00
+        0x00.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte(),
+        0x00.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte(),
+        0x00.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte(),
+        0x00.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte(),
+        0x00.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte()
     )
 }
